x86_64: Choose note information corredponding to proper task context
Note information for the current task context had wrongly been used for all
tasks in addition to the thread group leader. So, all the register values
contained in the generated core are equally the one for the thread group
leader. The correct logic is to choose the task context separately that is
passed to genregs_get() as @target argument.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@jp.fujitsu.com>