Add gcore_arch_get_gate_vma() and support VSYSCALL
Note that gcore_arch_get_gate_vma() is now only effective on X86_64. Also, any
resulting ELF core doesn't contain data in VSYSCALL page, since gate_vma's flag
on X86_64 is zero so dump filter mechanism fully excludes it. Looking at the
corresponding entry of a result of readelf -l helps to see this:

$ readelf -l core.3128
Program Headers:
  Type           Offset             VirtAddr           PhysAddr
                 FileSiz            MemSiz              Flags  Align
  NOTE           0x0000000000000468 0x0000000000000000 0x0000000000000000
                 0x0000000000000558 0x0000000000000000         0
  LOAD           0x0000000000001000 0x0000000000400000 0x0000000000000000
                 0x0000000000001000 0x0000000000006000  R E    1000
<snip>
  LOAD           0x000000000005b000 0xffffffffff600000 0x0000000000000000
                 0x0000000000000000 0x0000000000001000  R E    1000

The last entry corresponds to the one for gate_vma and its FileSiz is zero.
This means the memory content is not contained in segment.