Fix unintensionally dropping NT_386_IOPERM
Currently, NT_386_XSTATE note information has been dropped off since it has yet
to be tested. The problem is the corresponding index value, REGSET_XSTATE, is
in the middle of the values of the same enumeration type, and so note entry for
NT_386_IOPERM had not been written into coredump although it has been counted
and the resulting p_filesz in PT_NOTE has been the one that contained it. To
fix this, this patch moves temporarily NT_386_XSTATE's definition into the end
so that it no longer effects how NT_386_IOPERM is written into the coredump.