Revert regset get interface to the original
I simplified interface of regset get method so it returns TRUE if in
successful state and FALSE if in failure state. Original one returns 0
in successful state and error stats values in failure states. But the
reason of introducing regset interface is to make it easy to port
origianal regset, so it's not reasonable to change the interface.
(Still there's the opinion that error status values are now mostly
meaningless from gcore's point of view.)

Suggested-by: Lei Wen <leiwen@marvell.com>