gcore: add VM_DONTDUMP flag
VM_DONTDUMP flag was introduced in the following commit for user
applications to say we don't want this area in core dump. The typical
user is qemu which contains a large guest machine's memory.

    commit accb61fe7bb0f5c2a4102239e4981650f9048519
    Author: Jason Baron <jbaron@redhat.com>
    Date:   Fri Mar 23 15:02:51 2012 -0700

       coredump: add VM_NODUMP, MADV_NODUMP, MADV_CLEAR_NODUMP

To determine if VM_DONTDUMP flag is present, we look up vsyscall page
and check whether VM_ALWAYSDUMP flag is set on the vma corresponding
to the vsycall page.

The reason why I don't look up kernel version of a given dumpfile is
that this madvise flag has been backported to some distribution
kernels that are managed based on the kernels older than upstream
latest version.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@jp.fujitsu.com>