gcore: deal with removal of VM_RESERVED flag
The following patch around v3.6 removed VM_RESERVED flag.

    commit 314e51b9851b4f4e8ab302243ff5a6fc6147f379
    Author: Konstantin Khlebnikov <khlebnikov@openvz.org>
    Date:   Mon Oct 8 16:29:02 2012 -0700

        mm: kill vma flag VM_RESERVED and mm->reserved_vm counter

This patch deals with this by checking if mm_struct->reserved_vm
exists. Although we cannot look at VM_RESERVED flag due to the fact
that this is defined as macro, the above patch also removed this
member and we can use this as a mark to identify VM_RESERVED flag.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@jp.fujitsu.com>