coredump: Don't pass a list of threads to thread_group_cputime() method
We had passed a list of threads that belong to the same thread group
to thread_group_cputime() method explicitly since
thread_group_cputime_v22() needs to calculate total execution time of
all the threads in the same thread group.

However, we're going to stop creating a whole list of threads at the
same time to avoid memory pressure. In fact, we've just faced "cannot
allocate any more memory!" when we tried to create note information
for a whole list of threads.

In conclusion, we no longer pass a whole list. Instead, we iterate
task_context table managed by crash utility which includes all the
task_contexts in the vmcore. We restrict them by the unique thread
group ID.

Signed-of-by: HATAYAMA Daisuke <d.hatayama@jp.fujitsu.com>