gcore, x86: fix build failure on x86 due to no VDSO_HIGH_BASE definition
Currently, VDSO_HIGH_BASE is defined on x86_64 only. But unique
position where VDSO_HIGH_BASE is used is generic position, executed on
every architecture. I found this failure when building this on x86
architecture.

To address this issue, I remove restriction on x86_64 and export
VDSO_HIGH_BASE to every architecture; but actually used and should be
used only on x86_64 now.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@jp.fujitsu.com>