arm64, x86_64: Add ARM64 support
This patch was provided by Dave Anderson for a provision of upcoming
ARM64 support on RHEL7.1.

In addtition to ordinary ARM64 support, this patch also provides ARM64
on x86_64 binary support. In other words, if we build a crash utility
with TARGET=ARM64 on x86_64 arch system, we get a x86_64 crash utility
executable by which we can analize ARM64 crash dump.

Signed-off-by: Dave Anderson <anderson@redhat.com>
Signed-off-by: HATAYAMA Daisuke <d.hatayama@jp.fujitsu.com>