gcore, x86: put 0 in rsp register if no appropriate old_rsp location is found
If no appropriate old_rsp location is found, gxt->get_old_rsp is left
NULL, resulting in segmentation fault due to NULL pointer dereference.

This patch changes the behavior so that gcore puts 0 in that case to
make sure that at least a whole phase of gcore can be completed.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@jp.fujitsu.com>