gcore, coredump, x86: Add a helper function gcore_readmem_user()
As a preparation of ZRAM support, we introduce a new helper function
gcore_readmem_user() that integrates readswap(), which was introduced
to crash utility at 7.3.0, into a pair of uvtop() and readmem() to
avoid cluttering making the source code.

For crash utility before 7.3.0 where no readswap() is provided, define
readswap() in crash gcore command as a weak symbol.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@fujitsu.com>