Makefile: change testing build directory to our own local directory
We had used crash/extensions as build directory but it required crash
source tree and so we couldn't say it was handy enough. Instead, we
use our own local directory but then we still need defs.h so we keep
CRASH= command-line option to specify the directory where defs.h
is. Also, we modify clean: rule so that we remove src/defs.h and
transition into the gcore.mk's clean rule instead of into crash's
clean rule.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@fujitsu.com>