x86: fix how to identify 32-bit emulation
On the kernels v5.11 or later, gcore command generates core files in
the ELF64 format for the tasks running in 32-bit compatibility mode.

This is due to the change corresponding to the following commit on the
kernel:

    # git log -1 --stat 8d71d2bf6efec3032208958c483a247f529ffb16
    commit 8d71d2bf6efec3032208958c483a247f529ffb16
    Author: Gabriel Krisman Bertazi <krisman@collabora.com>
    Date:   Sat Oct 3 23:25:36 2020 -0400

	x86: Reclaim TIF_IA32 and TIF_X32

	Now that these flags are no longer used, reclaim those TIF bits.

	Signed-off-by: Gabriel Krisman Bertazi <krisman@collabora.com>
	Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
	Link: https://lore.kernel.org/r/20201004032536.1229030-11-krisman@collabora.com

     arch/x86/include/asm/thread_info.h | 4 ----
     arch/x86/kernel/process_64.c       | 6 ------
     2 files changed, 10 deletions(-)

where TIF_IA32 flag was removed. As the result, tasks running in
32-bit compatibility mode is never marked using TIF_IA32 flag.

To fix this issue, distinguish running mode by the value contained in
cs segment register just as user_64bit_mode().

This logic has been tested on RHEL6 (based on kernel v2.32), RHEL7
(based on kernel v3.10), RHEL8 (based on kernel v4.18), and Fedora 35
and RHEL9 beta (based on kernel v5.14), including the ones where
TIF_IA32 is still present.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@fujitsu.com>