Update Copyright notice in version information
Update Copyright notice in version information that is displayed when
gcore command is run with -V option.

The update of the notation on Copyright year is the same as those done
for Copyright notices in the respective files. On the other hand, the
notation of corporation name doesn't need to be changed because it has
already been Fujitsu Limited since it was first created.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@fujitsu.com>