x86: fix failure of loading gcore.so due to introduction of maple tree management on vma list
For memory dump files of Linux kernel v6.1 or later, loading gcore.so
fails with the following messages:

    crash> extend gcore.so

    extend: invalid structure member offset: mm_struct_mmap
            FILE: libgcore/gcore_x86.c  LINE: 2528  FUNCTION: gcore_arch_vsyscall_has_vm_alwaysdump_flag()

    [/root/repos/crash/crash] error trace: 7fd66441c820 => 7fd66441aca8 => 5d8f17 => 5d8ea3

      5d8ea3: OFFSET_verify.part.0+51
      5d8f17: OFFSET_verify+39

    extend: invalid structure member offset: mm_struct_mmap
            FILE: libgcore/gcore_x86.c  LINE: 2528  FUNCTION: gcore_arch_vsyscall_has_vm_alwaysdump_flag()

This is caused by the introduction of maple tree management on vma
list at Linux kernel v6.1, where mmap member of struct mm_struct was
removed, breaking the load-time processing of exploring the value of
VM_ALWAYSDUMP flag.

To fix this issue, consider that VM_ALWAYSDUMP flag was already not
present when maple tree was introduced. This is simple and sufficient.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@fujitsu.com>