coredump: fix failure of executing gcore command due to introduction of maple tree management on vma list
For memory dump files of Linux kernel v6.1 or later, executing gcore
command fails with the following messages:

    crash> gcore 1

    gcore: invalid structure member offset: mm_struct_mmap
           FILE: libgcore/gcore_coredump.c  LINE: 143  FUNCTION: gcore_coredump()

    [/root/repos/crash/crash] error trace: 7f5fd88ca6c2 => 7f5fd88bea09 => 5d8f17 => 5d8ea3

      5d8ea3: OFFSET_verify.part.0+51
      5d8f17: OFFSET_verify+39

    gcore: invalid structure member offset: mm_struct_mmap
           FILE: libgcore/gcore_coredump.c  LINE: 143  FUNCTION: gcore_coredump()

    Failed.

This is caused by the introduction of maple tree management on vma
list at Linux kernel v6.1. So far, gcore command gathers vma objects
of a given task via a linked-list of vma. However, the linked-list was
removed by the introduction of maple tree. As a result, gcore command
results in invalid reference to the non-existing mmap member of struct
mm_struct.

To address this issue, let's use maple tree API that has newly been
implemented at crash utility according to the introduction of maple
tree in Linux kernel. The maple tree API provides a utility function
do_maple_tree() to make several operations on maple tree. gcore
command uses the do_maple_tree() to collect a list of vma objects of a
given task.

Note that backward-compatibility for the old kernels including v6.0
and former is of course maintained.

This is implemented based on the implementation of process core
dumping feature of Linux kernel, borrowing data structures such as
struct core_vma_metadata. This policy comes from thought that keeping
the code base similar to the Linux kernel's increases maintainability.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@fujitsu.com>