coredump: Fix failure of executing gcore command due to memory write out of bounds.
executing gcore command fails with the following messages:
	crash> gcore -v 0 1
	free(): invalid next size (normal)
	Aborted

This is caused by the buffer size allocated for fill_files_note() is estimated as size = cprm->vma_count * 64, when filenames are generally long, it may lead to a write-out-of-bounds error, which occurs during free.

The modification involves increasing the estimated size or calculating the total length of filenames in all VMAs, and then adding metadata information to obtain the size. Here, the latter approach is adopted.